server/requirements.txt
flask==2.0.1
Flask-Cors==3.0.10
flask_restful==0.3.9
requests==2.25.1
# mysql
mysql-connector-python==8.0.12
SQLAlchemy==1.4.22

python-dotenv